+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - ./bin/generate_service_yaml.php to generate `services.yaml` for `oe-console`
   - ./bin/oxrun-light.php small application with one command `misc:register:command`
   - Command `misc:generate:yaml:config` has new Option `--list` to show all configrations
+  - Command `misc:generate:yaml:config` has new Options `--production` `--staging` `--development` `--testing`
 
 ### Changed
   - Oxrun is now a [OXID eShop Component](https://docs.oxid-esales.com/developer/en/6.2/development/modules_components_themes/component.html)